TypeRegistry
* Move TypeHelper functionality to TypeRegistry
* TypePropertyNotFoundException fixup
* Create Type registries for Reference, SQLite, PostgreSQL, MySQL
* Default number type flags
* add hard wired TypeInterface methods for more ease of use
* Integer type <-> column matching improvements
* Column / Type media type support
* MySQL type registry cleanup

+++ tests/reference/DBMSCommon_MySQL_create_types.sql
@@ -5,7 +5,7 @@ OR REPLACE TABLE `ns_unittests`.`types` (
   `boolean` boolean DEFAULT TRUE, 
   `int` bigint(20) AUTO_INCREMENT, 
   `large_int` bigint(12) DEFAULT 123456789012, 
-  `small_int` tinyint(3) UNSIGNED, 
+  `small_int` smallint(3) UNSIGNED, 
   `float` decimal(30, 2) DEFAULT 1.23, 
   `timestamp` datetime DEFAULT '2010-11-12 13:14:15', 
   `timestamp_tz` timestamp DEFAULT '2010-11-12 13:14:15',
